Update the CentOS 8 kickstart notes
The `repo` directives won't follow URL schemes starting with `file://`.
So, the more I think about it, the more I'm inclined to delete the
BaseOS and AppStream directories from the modified ISO file. If the
installer isn't going to read/use them, why keep them? So, anticipate
another set of changes in the next week.


- modified:   kickstart/centos8/README.md
- modified:   kickstart/centos8/isoBuilder.sh
- modified:   kickstart/centos8/ks.cfg

+++ kickstart/centos8/isoBuilder.sh
@@ -2,50 +2,87 @@
 set -x
 shopt -s dotglob
 
-df_source_iso=/home/drastic/Downloads/CentOS-8.3.2011-x86_64-dvd1.iso
+
+###
+### EXPLICIT VARIABLES
+###
+
+df_source_iso=/tmp/CentOS-8.3.2011-x86_64-dvd1.iso
 d_source_iso_mountpoint=/mnt
 
-df_kickstart=/run/media/drastic/flashDrive/ks.cfg
+df_kickstart=/tmp/custom-ks.cfg
 
-df_dest_iso=/home/CentOS-8.3-2011-x86_64-ks.iso
-d_build_dir=/home/iso
+df_dest_iso=/tmp/CentOS-8.3-2011-x86_64-ks.iso
+d_build_dir=/tmp/iso
 d_usb_device=/dev/sdb
 
 
-rm -rf ${df_dest_iso}
+###
+### DERIVED VARIABLES
+###
 
-mount -o loop ${df_source_iso} ${d_source_iso_mountpoint}
+s_source_label=$(blkid \
+                --match-tag=LABEL \
+                -o export \
+                ${df_source_iso} \
+                | grep LABEL \
+                | cut -d= -f2)
+
+
+###
+### MAIN
+###
+
+if [ -z ${s_source_label} ]
+then
+  exit 1
+fi
+
+if [ ! -e ${df_kickstart} ]
+then
+  exit 1
+fi
+
+ksvalidator ${df_kickstart}
+if [ $? != 0 ]
+then
+  exit 1
+fi
 
 if [ ! -d ${d_build_dir} ]
 then
   mkdir -p ${d_build_dir}
 fi
 
+
+rm -rf ${df_dest_iso}
+mount -o loop ${df_source_iso} ${d_source_iso_mountpoint}
 rm -rf ${d_build_dir}/*
 time cp -avRfp ${d_source_iso_mountpoint}/* ${d_build_dir}
-
 cp -vf ${df_kickstart} ${d_build_dir}/ks.cfg
 
 
 sed --in-place=.orig \
-  "/append/ s/$/ ks=cdrom:\/ks.cfg/" \
+  "/append/ s%$% inst.ks=hd:LABEL=${s_source_label}%" \
   ${d_build_dir}/isolinux/isolinux.cfg
 diff \
   ${d_build_dir}/isolinux/isolinux.cfg.orig \
   ${d_build_dir}/isolinux/isolinux.cfg
 
 sed --in-place=.orig \
-  "/linuxefi/ s/$/ inst.ks=cdrom:\/ks.cfg/" \
+  "/linuxefi/ s%$% inst.ks=hd:LABEL=${s_source_label}%" \
   ${d_build_dir}/EFI/BOOT/grub.cfg
 diff \
   ${d_build_dir}/EFI/BOOT/grub.cfg.orig \
   ${d_build_dir}/EFI/BOOT/grub.cfg
+
 
 cd ${d_build_dir}
 time mkisofs \
   -o ${df_dest_iso} \
   -b isolinux/isolinux.bin \
   -J -R -l -v \
+  -quiet \
   -c isolinux/boot.cat \
   -no-emul-boot \
   -boot-load-size 4 \
@@ -54,12 +91,11 @@ time mkisofs \
   -e images/efiboot.img \
   -no-emul-boot \
   -graft-points \
-  -V "CentOS-8-3-2011-x86_64-dvd" \
+  -V "${s_source_label}" \
   -jcharset utf-8 .
-
 isohybrid --uefi ${df_dest_iso}
-implantisomd5 ${df_dest_iso}
+time implantisomd5 ${df_dest_iso}
+
 
 umount ${d_usb_device}1
 time dd if=${df_dest_iso} of=${d_usb_device} status=progress
-
